# doc/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "NexusHub V1 API",
    "version": "1.0.0-design",
    "description": "开发前契约，尚无运行服务。全部业务POST，只有CSRF引导为GET。跨字段领域规则见模块文档。"
  },
  "servers": [
    {
      "url": "/",
      "description": "同源部署"
    }
  ],
...
  "paths": {
    "/api/auth/csrf": {
      "get": {
        "operationId": "getCsrf",
        "tags": [
          "identity"
        ],
        "summary": "获取CSRF令牌",
        "description": "权限：匿名。",
        "security": [],
        "responses": {
          "200": {
            "description": "成功（异步创建仅表示已入队）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CsrfResponse"
                }
              }
            }
          },
          "400": {
            "description": "参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "未认证",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "权限或CSRF错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "不存在或不可访问",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "版本、状态或幂等冲突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "规则校验失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "限流或队列已满",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "服务异常",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "operationId": "login",
        "tags": [
          "identity"
        ],
        "summary": "登录",
        "description": "权限：匿名。",
        "security": [
          {
            "CsrfHeader": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功（异步创建仅表示已入队）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "400": {
            "description": "参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "未认证",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "权限或CSRF错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "不存在或不可访问",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "版本、状态或幂等冲突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "规则校验失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "限流或队列已满",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "服务异常",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        }
      }
    },
    "/api/auth/me": {
      "post": {
        "operationId": "currentUser",
        "tags": [
          "identity"
        ],
        "summary": "当前用户",
        "description": "权限：登录。",
        "security": [
          {
            "SessionCookie": [],
            "CsrfHeader": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功（异步创建仅表示已入队）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "400": {
            "description": "参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "未认证",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "权限或CSRF错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "不存在或不可访问",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "版本、状态或幂等冲突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "规则校验失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "限流或队列已满",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "服务异常",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Empty"
              }
            }
          }
        }
      }
    },
    "/api/auth/guest": {
      "post": {
        "operationId": "createGuest",
        "tags": [
          "identity"
        ],
        "summary": "创建游客会话",
        "description": "权限：匿名 + CSRF。",
        "security": [
          {
            "CsrfHeader": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功（异步创建仅表示已入队）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "400": {
            "description": "参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "未认证",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "权限或CSRF错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "不存在或不可访问",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "版本、状态或幂等冲突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "规则校验失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "限流或队列已满",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "服务异常",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Empty"
              }
            }
          }
        }
      }
    },
    "/api/auth/logout": {
      "post": {
        "operationId": "logout",
        "tags": [
          "identity"
        ],
        "summary": "退出",
        "description": "权限：登录。",
        "security": [
          {
            "SessionCookie": [],
            "CsrfHeader": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功（异步创建仅表示已入队）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmptyResponse"
                }
              }
            }
          },
          "400": {
            "description": "参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "未认证",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "权限或CSRF错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "不存在或不可访问",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "版本、状态或幂等冲突",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "规则校验失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "限流或队列已满",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "服务异常",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Empty"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "SessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "NEXUSHUB_SESSION"
      },
      "CsrfHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-TOKEN"
      }
    },
    "schemas": {
...
      "Empty": {
        "type": "object",
        "additionalProperties": false,
        "properties": {},
        "required": []
      },
...
      "FieldError": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "field": {
            "type": "string",
            "maxLength": 128
          },
          "reason": {
            "type": "string",
            "maxLength": 300
          }
        },
        "required": [
          "field",
          "reason"
        ]
      },
      "ErrorData": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "fieldErrors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FieldError"
            },
            "minItems": 0,
            "maxItems": 100
          },
          "currentVersion": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0,
                "maximum": 2147483647
              },
              {
                "type": "null"
              }
            ]
          },
          "retryAfterSeconds": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0,
                "maximum": 2147483647
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": []
      },
      "ErrorResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "code": {
            "type": "integer",
            "minimum": 40000,
            "maximum": 59999
          },
          "message": {
            "type": "string",
            "maxLength": 500
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorData"
              },
              {
                "type": "null"
              }
            ]
          },
          "requestId": {
            "type": "string",
            "maxLength": 64
          }
        },
        "required": [
          "code",
          "message",
          "data",
          "requestId"
        ]
      },
      "Csrf": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "token": {
            "type": "string",
            "maxLength": 512
          },
          "headerName": {
            "type": "string",
            "enum": [
              "X-CSRF-TOKEN"
            ]
          }
        },
        "required": [
          "token",
          "headerName"
        ]
      },
      "LoginRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "username": {
            "type": "string",
            "maxLength": 64,
            "minLength": 3
          },
          "password": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "format": "password"
          }
        },
        "required": [
          "username",
          "password"
        ]
      },
      "User": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "maxLength": 19,
            "pattern": "^[1-9][0-9]{0,18}$",
            "description": "正整数 BIGINT 的十进制字符串，服务端还校验 <= 9223372036854775807"
          },
          "username": {
            "type": "string",
            "maxLength": 64
          },
          "displayName": {
            "type": "string",
            "maxLength": 64
          },
          "role": {
            "type": "string",
            "enum": [
              "GUEST",
              "USER",
              "ADMIN"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "ENABLED",
              "DISABLED"
            ]
          },
          "rowVersion": {
            "type": "integer",
            "minimum": 0,
            "maximum": 2147483647
          },
          "permissions": {
            "type": "array",
            "items": {
              "type": "string",
              "maxLength": 64
            },
            "minItems": 0,
            "maxItems": 100
          }
        },
        "required": [
          "id",
          "username",
          "displayName",
          "role",
          "status",
          "rowVersion",
          "permissions"
        ]
      },
...
      "CsrfResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "code": {
            "const": 200,
            "type": "integer"
          },
          "message": {
            "type": "string",
            "maxLength": 500
          },
          "data": {
            "$ref": "#/components/schemas/Csrf"
          },
          "requestId": {
            "type": "string",
            "maxLength": 64
          }
        },
        "required": [
          "code",
          "message",
          "data",
          "requestId"
        ]
      },
      "UserResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "code": {
            "const": 200,
            "type": "integer"
          },
          "message": {
            "type": "string",
            "maxLength": 500
          },
          "data": {
            "$ref": "#/components/schemas/User"
          },
          "requestId": {
            "type": "string",
            "maxLength": 64
          }
        },
        "required": [
          "code",
          "message",
          "data",
          "requestId"
        ]
      },
      "EmptyResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "code": {
            "const": 200,
            "type": "integer"
          },
          "message": {
            "type": "string",
            "maxLength": 500
          },
          "data": {
            "$ref": "#/components/schemas/Empty"
          },
          "requestId": {
            "type": "string",
            "maxLength": 64
          }
        },
        "required": [
          "code",
          "message",
          "data",
          "requestId"
        ]
      },
...
    }
  }
}